Update AndroidManifest.xml
Public activities can be used by an unspecified large number of applications, including malware. When using public activities it is also important to note that malware can read and receive any intents sent to them. This activity is not properly protected from malware.

Private activities cannot be launched by the other applications and therefore it is the safest activity. When addressing private activities, it is recommended to use explicit intents. This will prevent accidentally sending the intent to other applications, such as malware, that are capable of handling.